+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Commands/ShowMaintenanceCommand.cs
@@ -1,5 +1,5 @@
 using System.Windows;
-using TD.WPF.Views;
+using TD.WPF.Views.Maintenance;
 
 namespace TD.WPF.Commands;