+++ OzzTradeDiary.NET/OzzTradeDiary.WPF/Views/MainWindow.xaml.cs
@@ -1,18 +1,8 @@
-﻿using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using TD.Models;
+﻿using System.Windows;
 using TD.WPF.Models;
 using TD.WPF.ViewModels;
 
-namespace TD.WPF
+namespace TD.WPF.Views
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml